Create TokenManagerERC20.sol
Add function exitToMainRecipientERC20. This transfers tokens from msg.sender on the schain to a specified recipient address on mainnet.
For checkAllowedToSendMessage in exitToMainRecipientERC20, we use tx.origin, not msg.sender. This ensures the rate limiting is on the user account, not the proxy contract.
Add function transferToSchainRecipientERC20. This transfers tokens from msg.sender on the schain to a specified recipient address on another schain.